requirements.txt
roboflow
opencv-python
pyserial